+++ R/plot_amt.R
@@ -15,21 +15,21 @@
 #' 
 #' @export
 amt_vs_idv <- function(xpdb,
-                       mapping  = NULL,
-                       group    = 'ID',
+                       mapping    = NULL,
+                       group      = 'ID',
                        drop_fixed = TRUE,
-                       type     = 'l',
-                       title    = 'Compartments amount vs. @x | @run',
-                       subtitle = 'Ofv: @ofv',
-                       caption  = '@dir',
-                       log      = NULL,
+                       type       = 'l',
+                       title      = 'Compartments amount vs. @x | @run',
+                       subtitle   = 'Ofv: @ofv',
+                       caption    = '@dir',
+                       log        = NULL,
                        facets,
-                       problem,
+                       .problem,
                        quiet,
                        ...) {
   # Check input
   check_xpdb(xpdb, check = 'data')
-  if (missing(problem)) problem <- default_plot_problem(xpdb)
+  if (missing(.problem)) .problem <- default_plot_problem(xpdb)
   if (missing(quiet)) quiet <- xpdb$options$quiet
   if (missing(facets)) facets <- add_facet_var(facets = xpdb$xp_theme$facets, 
                                                variable = 'variable')
@@ -38,9 +38,9 @@ amt_vs_idv <- function(xpdb,
   if (!any(names(extra_args) == 'nrow')) extra_args$nrow <- 3
   if (!any(names(extra_args) == 'ncol')) extra_args$ncol <- 3
   
-  amt_col <- xp_var(xpdb, problem, type = 'a')$col
+  amt_col <- xp_var(xpdb, .problem, type = 'a')$col
   if (drop_fixed) {
-    amt_col <- drop_fixed_cols(xpdb, problem, cols = amt_col, quiet = quiet)
+    amt_col <- drop_fixed_cols(xpdb, .problem, cols = amt_col, quiet = quiet)
   }
   if (is.null(amt_col)) {
     stop('No compartment amount column found in the xpdb data index.', call. = FALSE)
@@ -49,12 +49,12 @@ amt_vs_idv <- function(xpdb,
   do.call('xplot_scatter', 
           c(extra_args, 
             list(xpdb = xpdb, group = group, quiet = quiet,
-                 opt = data_opt(problem = problem, 
+                 opt = data_opt(.problem = .problem, 
                                 filter = function(x) {
                                   dplyr::select_if(.tbl = x, .predicate = function(x) dplyr::n_distinct(x) > 1)
                                 }, tidy = TRUE, value_col = amt_col,
                                 post_processing = reorder_factors(prefix = 'Comp. ')),
-                 mapping = aes_c(aes_string(x = xp_var(xpdb, problem, type = 'idv')$col, 
+                 mapping = aes_c(aes_string(x = xp_var(xpdb, .problem, type = 'idv')$col, 
                                             y = 'value'), mapping),
                  type = type, facets = facets, 
                  xscale = check_scales('x', log),