+++ R/vars_set.R
@@ -3,7 +3,7 @@
 #' @description Function designed to change the type, label or unit associated with variables.
 #' 
 #' @param xpdb An \code{xpose_data} object.
-#' @param problem The problem number to which the edits will be applied.
+#' @param .problem The problem number to which the edits will be applied.
 #' @param ... Specifications of the edits to be made to the xpdb index. Edits are made as 
 #' type and variable pairs e.g. idv = 'TAD' will assign TAD to the type idv (independent variable).
 #' @param auto_factor With \code{set_var_types} only. If \code{TRUE} new columns assigned to the type 'catcov' will be converted to
@@ -35,28 +35,28 @@
 #' @seealso \code{\link{list_vars}}
 #' @examples
 #' # Change variable type
-#' xpdb_2 <- set_var_types(xpdb_ex_pk, problem = 1, idv = 'TAD')
+#' xpdb_2 <- set_var_types(xpdb_ex_pk, .problem = 1, idv = 'TAD')
 #' 
 #' # Change labels
-#' xpdb_2 <- set_var_labels(xpdb_2, problem = 1, ALAG1 = 'Lag time', CL = 'Clearance', V = 'Volume')
+#' xpdb_2 <- set_var_labels(xpdb_2, .problem = 1, ALAG1 = 'Lag time', CL = 'Clearance', V = 'Volume')
 #' 
 #' # Change units
-#' xpdb_2 <- set_var_units(xpdb_2, problem = 1, ALAG1 = 'h', CL = 'L/h', V = 'L')
+#' xpdb_2 <- set_var_units(xpdb_2, .problem = 1, ALAG1 = 'h', CL = 'L/h', V = 'L')
 #' 
 #' @name set_vars
 #' @export
-set_var_types <- function(xpdb, problem = NULL, ..., auto_factor = TRUE, quiet) {
+set_var_types <- function(xpdb, .problem = NULL, ..., auto_factor = TRUE, quiet) {
   # Check input
   check_xpdb(xpdb, check = 'data')
   if (missing(quiet)) quiet <- xpdb$options$quiet
   
   dat <- xpdb$data
   
-  if (!is.null(problem) && !all(problem %in% dat$problem)) {
-    stop('Problem no.', stringr::str_c(problem[!problem %in% dat$problem], collapse = ', '), 
+  if (!is.null(.problem) && !all(.problem %in% dat$problem)) {
+    stop('Problem no.', stringr::str_c(.problem[!.problem %in% dat$problem], collapse = ', '), 
          ' not found in model output data.', call. = FALSE)
   }
-  if (is.null(problem)) problem <- unique(dat$problem)
+  if (is.null(.problem)) .problem <- unique(dat$problem)
   
   args <- c(...)
   if (is.null(args)) return(xpdb)
@@ -69,7 +69,7 @@ set_var_types <- function(xpdb, problem = NULL, ..., auto_factor = TRUE, quiet) 
     dplyr::mutate(grouping = .$problem) %>% 
     dplyr::group_by_(.dots = 'grouping') %>% 
     tidyr::nest(.key = 'tmp') %>% 
-    dplyr::mutate(out = purrr::map_if(.$tmp, .$grouping %in% problem, function(x, args, quiet) {
+    dplyr::mutate(out = purrr::map_if(.$tmp, .$grouping %in% .problem, function(x, args, quiet) {
       # Get the index
       index <- x$index[[1]]
       
@@ -109,18 +109,18 @@ set_var_types <- function(xpdb, problem = NULL, ..., auto_factor = TRUE, quiet) 
 }
 
 
-set_var_generic <- function(xpdb, problem = NULL, what = NULL, ..., quiet) {
+set_var_generic <- function(xpdb, .problem = NULL, what = NULL, ..., quiet) {
   # Check input
   check_xpdb(xpdb, check = 'data')
   if (missing(quiet)) quiet <- xpdb$options$quiet
   
   dat <- xpdb$data
   
-  if (!is.null(problem) && !all(problem %in% dat$problem)) {
-    stop('Problem no.', stringr::str_c(problem[!problem %in% dat$problem], collapse = ', '), 
+  if (!is.null(.problem) && !all(.problem %in% dat$problem)) {
+    stop('Problem no.', stringr::str_c(.problem[!.problem %in% dat$problem], collapse = ', '), 
          ' not found in model output data.', call. = FALSE)
   }
-  if (is.null(problem)) problem <- unique(dat$problem)
+  if (is.null(.problem)) .problem <- unique(dat$problem)
   
   args <- c(...)
   if (is.null(args)) return(xpdb)
@@ -131,7 +131,7 @@ set_var_generic <- function(xpdb, problem = NULL, what = NULL, ..., quiet) {
     dplyr::mutate(grouping = .$problem) %>% 
     dplyr::group_by_(.dots = 'grouping') %>% 
     tidyr::nest(.key = 'tmp') %>% 
-    dplyr::mutate(out = purrr::map_if(.$tmp, .$grouping %in% problem, function(x, args, quiet) {
+    dplyr::mutate(out = purrr::map_if(.$tmp, .$grouping %in% .problem, function(x, args, quiet) {
       # Get the index
       index <- x$index[[1]]
       
@@ -158,14 +158,14 @@ set_var_generic <- function(xpdb, problem = NULL, what = NULL, ..., quiet) {
 
 #' @rdname set_vars
 #' @export
-set_var_labels <- function(xpdb, problem = NULL, ..., quiet) {
-  set_var_generic(xpdb = xpdb, problem = problem, 
+set_var_labels <- function(xpdb, .problem = NULL, ..., quiet) {
+  set_var_generic(xpdb = xpdb, .problem = .problem, 
                   quiet = quiet, what = 'label', ...)
 }
 
 #' @rdname set_vars
 #' @export
-set_var_units <- function(xpdb, problem = NULL, ..., quiet) {
-  set_var_generic(xpdb = xpdb, problem = problem, 
+set_var_units <- function(xpdb, .problem = NULL, ..., quiet) {
+  set_var_generic(xpdb = xpdb, .problem = .problem, 
                   quiet = quiet, what = 'units', ...)
 }