+++ tests/testthat/test-fetch_data.R
@@ -2,7 +2,7 @@ context('Check fetch_data')
 
 # Tests start here --------------------------------------------------------
 test_that('data_opt function works properly', {
-  expect_equal(data_opt(problem = 1, subprob = 2, method = 'foce', source = 'data', 
+  expect_equal(data_opt(.problem = 1, .subprob = 2, .method = 'foce', .source = 'data', 
                         simtab = FALSE, filter = function(x){ x }, tidy = FALSE, index_col = 'ID',
                         value_col = 'IPRED', post_processing = function(x) { x }),
                list(problem = 1, subprob = 2, method = 'foce', source = 'data', simtab = FALSE, 
@@ -11,14 +11,14 @@ test_that('data_opt function works properly', {
 })
 
 test_that('only_obs function works properly', {
-  expect_true(is.function(only_obs(xpdb_ex_pk, problem = 1, quiet = TRUE)))
-  expect_equal(only_obs(xpdb_ex_pk, problem = 1, quiet = TRUE)(x = xpdb_ex_pk$data$data[[1]]),
+  expect_true(is.function(only_obs(xpdb_ex_pk, .problem = 1, quiet = TRUE)))
+  expect_equal(only_obs(xpdb_ex_pk, .problem = 1, quiet = TRUE)(x = xpdb_ex_pk$data$data[[1]]),
                xpdb_ex_pk$data$data[[1]][xpdb_ex_pk$data$data[[1]]$EVID == 0, ])
 })
 
 test_that('only_distinct function works properly', {
-  expect_true(is.function(only_distinct(xpdb_ex_pk, problem = 1, facets = 'OCC', quiet = TRUE)))
-  expect_equal(only_distinct(xpdb_ex_pk, problem = 1, facets = 'OCC', quiet = TRUE)(x = xpdb_ex_pk$data$data[[1]]),
+  expect_true(is.function(only_distinct(xpdb_ex_pk, .problem = 1, facets = 'OCC', quiet = TRUE)))
+  expect_equal(only_distinct(xpdb_ex_pk, .problem = 1, facets = 'OCC', quiet = TRUE)(x = xpdb_ex_pk$data$data[[1]]),
                dplyr::distinct_(.data = xpdb_ex_pk$data$data[[1]], .dots = c('ID', 'OCC'), .keep_all = TRUE))
 })
 
@@ -27,7 +27,7 @@ test_that('reorder_etas_factors function works properly', {
 })
 
 test_that('fetch_data can get simple data', {
-  imported_data <- fetch_data(xpdb_ex_pk, problem = 1, source = 'data', 
+  imported_data <- fetch_data(xpdb_ex_pk, .problem = 1, .source = 'data', 
                               simtab = FALSE, filter = only_obs(xpdb_ex_pk, 1, TRUE))
   
   # Check attributes
@@ -40,7 +40,7 @@ test_that('fetch_data can get simple data', {
 })
 
 test_that('fetch_data can tidy data', {
-  imported_tidy_data <- fetch_data(xpdb_ex_pk, problem = 1, source = 'data', 
+  imported_tidy_data <- fetch_data(xpdb_ex_pk, .problem = 1, .source = 'data', 
                                    simtab = FALSE, filter = function(x) {
                                      x[x[, 'EVID'] == 0, c('ID', 'TIME', 'IPRED', 'PRED', 'DV')]},
                                    tidy = TRUE, value_col = c('IPRED','PRED', 'DV'))
@@ -55,6 +55,6 @@ test_that('fetch_data can tidy data', {
 })
 
 test_that('fetch_data can get file data', {
-  expect_equal(fetch_data(xpdb_ex_pk, problem = 1, source = 'ext', quiet = TRUE), 
+  expect_equal(fetch_data(xpdb_ex_pk, .problem = 1, .source = 'ext', quiet = TRUE), 
                xpdb_ex_pk$files[xpdb_ex_pk$files$name == 'run001.ext',]$data[[1]])
 })